fix(storage): address files with 64-bit byte offsets
Page IDs stay 32-bit ints. File positions and lengths are now long so
pid * 4096 no longer overflows at 2 GiB.

+++ src/buffer/BufferManager.java
@@ -442,7 +442,7 @@ public class BufferManager {
 	private void writePageToDisk(String fileId, Page page) throws IOException {
 		writeIOCount.increment();
 		try (RandomAccessFile raf = new RandomAccessFile(fileId, "rw")) {
-			int offset = RawPage.getOffset(page.getPid());
+			long offset = RawPage.getOffset(page.getPid());
 			raf.seek(offset);
 			raf.write(page.getByteArray());
 		}
